+++ lib/json_schema_form.rb
@@ -1,4 +1,7 @@
 require_relative "json_schema_form/version"
 require_relative 'json_schema_form/type'
 require_relative 'json_schema_form/field'
+require_relative 'json_schema_form/response'
+require_relative 'json_schema_form/response_set'
+require_relative 'json_schema_form/form'
 require_relative 'json_schema_form/document'
