+++ lib/json_schema_form/field.rb
@@ -1,7 +1,6 @@
 require_relative 'field/field_methods'
 require_relative 'field/checkbox'
 require_relative 'field/date_input'
-require_relative 'field/form'
 require_relative 'field/header'
 require_relative 'field/info'
 require_relative 'field/number_input'
@@ -10,5 +9,3 @@ require_relative 'field/slider'
 require_relative 'field/static'
 require_relative 'field/switch'
 require_relative 'field/text_input'
-require_relative 'field/response'
-require_relative 'field/response_set'
