+++ lib/json_schema_form/type/base.rb
@@ -70,7 +70,6 @@ module JsonSchemaForm
           optional(:description).maybe(:string)
           optional(:default)
           optional(:examples)
-          optional(:max_score).maybe(:integer)
         end
       end
